backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==2.9.0
openai==1.10.0
scikit-learn==1.4.0
numpy==1.26.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
requests==2.31.0
aiofiles==23.2.1